fastapi/pyproject.toml
[tool.poetry]
name = "fastapi-app"
version = "0.1.0"
description = ""
authors = ["Sarthak <sarthaksom10@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
streamlit-option-menu = "^0.4.0"
openai = "^1.52.0"
pymssql = "^2.3.1"
openpyxl = "^3.1.5"
pandas = "^2.2.3"
python-docx = "^1.1.2"
python-pptx = "^1.0.2"
datasets = "^3.0.1"
plotly = "^5.24.1"
streamlit-plotly-events = "^0.0.6"
uvicorn = "^0.32.0"
pydantic = {extras = ["email"], version = "^2.9.2"}
fastapi = "^0.115.2"
pymysql = "^1.1.1"
bcrypt = "^4.2.0"
jwt = "^1.3.1"
datetime = "^5.5"
python-multipart = "^0.0.12"
logging = "^0.4.9.6"
google-cloud = "^0.34.0"
google-cloud-storage = "^2.18.2"
pyjwt = "2.7.0"
pathlib = "^1.0.1"
boto3 = "^1.35.48"
google-cloud-vision = "^3.8.0"
pymupdf = "^1.24.12"
llama-parse = "^0.5.12"
llama-index-core = "^0.11.20"
pillow = "<11.0.0"
nest-asyncio = "^1.6.0"
numpy = "1.26.3"
pinecone-client = "^5.0.1"
langchain = "^0.3.4"
langchain-text-splitters = "^0.3.0"
tqdm = "^4.66.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
